refactor: Remove ?? operators so that compiler will check FFI modules
From natefaubion on the PureScript Discord:

The compiler only does any validation if it can actually parse the ffi file. What can be parsed is specific to the JS parser lib that we use (and I believe it's currently pinned). So if your JS is quite conservative, it's probably fine, but once you start using newer stuff, or something like JSX, it won't be able to parse it. We felt it was better to treat this validation as an optimistic thing rather than keep restricting users in what JS they could write arbitrarily based on an unmaintained JS parser's feature set.